requirements.txt
# requirements.txt
Flask==3.0.0
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
sqlalchemy
psycopg2-binary  # For PostgreSQL support